+++ Sources/Onboarding/Models/WelcomeScreen.swift
@@ -10,3 +10,37 @@ import SwiftUI
 public enum WelcomeScreen {
     case apple(AppleWelcomeScreen.Configuration)
 }
+
+public extension WelcomeScreen {
+    static func apple(
+        accentColor: Color = .blue,
+        appDisplayName: String,
+        appIcon: Image,
+        features: [FeatureInfo],
+        privacyPolicyURL: URL? = nil,
+        titleSectionAlignment: HorizontalAlignment = .leading
+    ) -> Self {
+        .apple(
+            .init(
+                accentColor: accentColor,
+                appDisplayName: appDisplayName,
+                appIcon: appIcon,
+                features: features,
+                privacyPolicyURL: privacyPolicyURL,
+                titleSectionAlignment: titleSectionAlignment
+            )
+        )
+    }
+
+    func with(continueAction: @escaping () -> Void) -> Self {
+        switch self {
+        case let .apple(configuration):
+            return .apple(configuration.with(continueAction: continueAction))
+        }
+    }
+}
+
+@MainActor
+public extension WelcomeScreen {
+    static let mock: Self = .apple(.mock)
+}